Core: manage child loggers using weak_ptr's
Make constructors private in favor of ::create() factory method.

Fixes multithreading issues with accessing already destroyed logger.
Since we have atomic locking, it can happen (and happens in tests)
that particular logger is destroyed in another thread.

Also had to switch to recursive version of update_propagated_sinks():
this way it is guaranteed that loggers from the whole hierarchy are
preserved during the traversal.

Also allows to keep destructor empty.

Signed-off-by: Pavel Artsishevsky <polter.rnd@gmail.com>